cys confirm 완료

+++ WebContent/WEB-INF/views/detailWriter.jsp
@@ -36,7 +36,7 @@
 			<a
 				href="${pageContext.servletContext.contextPath }/board/modifyDetail.do?no=${detail.NO}">수정</a>
 			<a
-				href="${pageContext.servletContext.contextPath }/board/deleteDetail.do?no=${detail.NO}">삭제</a>
+				href="${pageContext.servletContext.contextPath }/board/deleteDetail.do?no=${detail.NO}" class="confirm">삭제</a>
 
 		</div>
 
@@ -122,6 +122,17 @@
 			<p class="mb-1">&copy; 2018-2019 재미나</p>
 		</footer>
 	</div>
+	
+	<script type="text/javascript">
+		var conf = document.getElementsByClassName("confirm");
+		var confirmIt = function (e) {
+			if(!confirm("정말 삭제 하시겠습니까?")) 
+				e.preventDefault();
+		};
+		for (var i = 0, l = conf.length; i < l; i++) {
+	        conf[i].addEventListener('click', confirmIt, false);
+	    };
+	</script>
 	<!-- Bootstrap core JavaScript
     ================================================== -->
 	<!-- Placed at the end of the document so the pages load faster -->
